#914: Adds tests for serializers for searchable objects
