+++ src/main/scala/no/ndla/searchapi/model/search/SearchableArticle.scala
@@ -7,8 +7,10 @@
 
 package no.ndla.searchapi.model.search
 
-import no.ndla.searchapi.model.domain.SearchableTaxonomyContext
-import org.joda.time.DateTime
+import java.util.TimeZone
+import org.joda.time.{DateTime, DateTimeZone}
+import org.json4s.{CustomSerializer, Extraction, Formats}
+import org.json4s.JsonAST.{JField, JObject}
 
 case class SearchableArticle(
     id: Long,
@@ -27,6 +29,53 @@ case class SearchableArticle(
     supportedLanguages: List[String],
     contexts: List[SearchableTaxonomyContext]
 )
+
+class SearchableArticleSerializer
+    extends CustomSerializer[SearchableArticle](_ =>
+      ({
+        case obj: JObject =>
+          implicit val formats: Formats = SearchableLanguageFormats.JSonFormats
+
+          val time = (obj \ "lastUpdated").extract[DateTime]
+          val tz = TimeZone.getDefault
+          val lastUpdated = new DateTime(time, DateTimeZone.forID(tz.getID))
+
+          SearchableArticle(
+            id = (obj \ "id").extract[Long],
+            title = SearchableLanguageValues("title", obj),
+            content = SearchableLanguageValues("content", obj),
+            visualElement = SearchableLanguageValues("visualElement", obj),
+            introduction = SearchableLanguageValues("introduction", obj),
+            metaDescription = SearchableLanguageValues("metaDescription", obj),
+            tags = SearchableLanguageList("tags", obj),
+            lastUpdated = lastUpdated,
+            license = (obj \ "license").extract[String],
+            authors = (obj \ "authors").extract[List[String]],
+            articleType = (obj \ "articleType").extract[String],
+            defaultTitle = (obj \ "defaultTitle").extract[Option[String]],
+            metaImageId = (obj \ "metaImageId").extract[Option[String]],
+            supportedLanguages =
+              (obj \ "supportedLanguages").extract[List[String]],
+            contexts =
+              (obj \ "contexts").extract[List[SearchableTaxonomyContext]]
+          )
+      }, {
+        case article: SearchableArticle =>
+          implicit val formats: Formats = SearchableLanguageFormats.JSonFormats
+          val languageFields: List[JField] =
+            List(
+              article.title.toJsonField("title"),
+              article.content.toJsonField("content"),
+              article.visualElement.toJsonField("visualElement"),
+              article.introduction.toJsonField("introduction"),
+              article.metaDescription.toJsonField("metaDescription"),
+              article.tags.toJsonField("tags")
+            ).flatten
+
+          val partialSearchableArticle = LanguagelessSearchableArticle(article)
+          val partialJObject = Extraction.decompose(partialSearchableArticle)
+          partialJObject.merge(JObject(languageFields: _*))
+      }))
 
 object LanguagelessSearchableArticle {
   case class LanguagelessSearchableArticle(
