+++ src/main/scala/no/ndla/searchapi/model/search/SearchableLearningPath.scala
@@ -7,9 +7,11 @@
 
 package no.ndla.searchapi.model.search
 
+import java.util.TimeZone
 import no.ndla.searchapi.model.api.learningpath.Copyright
-import no.ndla.searchapi.model.domain.SearchableTaxonomyContext
-import org.joda.time.DateTime
+import org.joda.time.{DateTime, DateTimeZone}
+import org.json4s.{CustomSerializer, Extraction, Formats}
+import org.json4s.JsonAST.{JField, JObject}
 
 case class SearchableLearningPath(id: Long,
                                   title: SearchableLanguageValues,
@@ -25,6 +27,53 @@ case class SearchableLearningPath(id: Long,
                                   license: Copyright,
                                   isBasedOn: Option[Long],
                                   contexts: List[SearchableTaxonomyContext])
+
+class SearchableLearningPathSerializer
+    extends CustomSerializer[SearchableLearningPath](_ =>
+      ({
+        case obj: JObject =>
+          implicit val formats: Formats = SearchableLanguageFormats.JSonFormats
+
+          val time = (obj \ "lastUpdated").extract[DateTime]
+          val tz = TimeZone.getDefault
+          val lastUpdated = new DateTime(time, DateTimeZone.forID(tz.getID))
+
+          SearchableLearningPath(
+            id = (obj \ "id").extract[Long],
+            title = SearchableLanguageValues("title", obj),
+            description = SearchableLanguageValues("description", obj),
+            coverPhotoId = (obj \ "coverPhotoUrl").extract[Option[String]],
+            duration = (obj \ "duration").extract[Option[Int]],
+            status = (obj \ "status").extract[String],
+            verificationStatus = (obj \ "verificationStatus").extract[String],
+            lastUpdated = lastUpdated,
+            defaultTitle = (obj \ "defaultTitle").extract[Option[String]],
+            tags = SearchableLanguageList("tags", obj),
+            learningsteps =
+              (obj \ "learningsteps").extract[List[SearchableLearningStep]],
+            license = (obj \ "license").extract[Copyright],
+            isBasedOn = (obj \ "isBasedOn").extract[Option[Long]],
+            contexts =
+              (obj \ "contexts").extract[List[SearchableTaxonomyContext]]
+          )
+      }, {
+        case lp: SearchableLearningPath =>
+          implicit val formats: Formats = SearchableLanguageFormats.JSonFormats
+
+          val languageFields: List[JField] =
+            List(
+              lp.title.toJsonField("title"),
+              lp.description.toJsonField("description"),
+              lp.tags.toJsonField("tags")
+            ).flatten
+
+          val partialSearchableLearningPath =
+            LanguagelessSearchableLearningPath(lp)
+          val partialJObject =
+            Extraction.decompose(partialSearchableLearningPath)
+          partialJObject.merge(JObject(languageFields: _*))
+
+      }))
 
 object LanguagelessSearchableLearningPath {
   case class LanguagelessSearchableLearningPath(
@@ -56,5 +105,4 @@ object LanguagelessSearchableLearningPath {
       searchableLearningPath.contexts
     )
   }
-
 }
